Поправил консольное отображения во всех функциях, вынес вспомогательную функцию make_cute_line в отдельный файл
